requirements.txt
Django
django-ckeditor
django-extensions
Werkzeug
